requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
duckduckgo-search
slowapi
pydantic
requests
